test(qa): add QA tests for inter-cluster messaging
Adds a container based test for encrypted inter-cluster messaging. The
test ensures a cluster of 2 brokers and 1 gateway and find each other
and form a cluster properly, and will verify that each endpoint (the
command API, the internal API, and the gateway's cluster API) are all
secured with the expected certificate.